mudando de float para double a variavel "nivel" no exercicio04-Rio

+++ 4-semestre/programacao-orientada-a-objetos/listas/lista01/src/main/java/br/edu/fatecfranca/exercicio04/Rio.java
@@ -5,7 +5,7 @@ public class Rio {
     public double nivel;
     public boolean poluido;
 
-    public Rio(String nome, float nivel, boolean poluido) {
+    public Rio(String nome, double nivel, boolean poluido) {
         this.nome = nome;
         this.nivel = nivel;
         this.poluido = poluido;
